rename module files to conform with PEP-8
