+++ src/rabbitmq/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-from .RabbitSpider import RabbitSpider
+from .rabbit_spider import RabbitSpider
